+++ conf/evolutions/default/2.sql
@@ -1,0 +1,9 @@
+# -- Ups
+
+ALTER TABLE payments ALTER COLUMN units type BIGINT;
+
+
+
+# -- Downs
+
+ALTER TABLE payments ALTER COLUMN units type NUMERIC;